add page tools ,add valid tools, add logback

+++ src/main/java/com/sai/core/utils/PageMapper.java
@@ -1,7 +1,6 @@
 package com.sai.core.utils;
 
-import com.sai.core.dto.ResultCode;
-
+import java.math.BigDecimal;
 import java.util.List;
 import java.util.Map;
 
@@ -9,5 +8,5 @@ public interface PageMapper<T> {
     List select4Page(Map<String, Object> param);
     int insert(T addInfo);
     int update(T updateInfo);
-    T getById(Long id);
+    T getById(BigDecimal id);
 }
